Remove term proximity ranking (#20)
* minor ranking tweaks

* Remove term proximity ranking.
Term proximity has been saved to a seperate branch for the future, but will not be used during ranking.
Experimenting with 10 warc files seemed to indicate, that the term proximity ranking actually worsened the
search results quite substantially. We might want to re-introduce something like it in the future, but will
probably have to devote more time into paramater-tuning. Maybe doing something naive like searching for bi-grams and tri-grams with bm25 ranking is sufficient?