Use fontsource for loading Montserrat (#40)
This commit removes the Montserrat font stored in public and replaces
them with the fonts installed with @fontsource/montserrat.

This also streamlines the font import process, and ensures that correct
typeface formats are loaded.

One thing to consider, is using variable fonts instead. I failed to get
loading with this setup however, but it seems to be supported.