Use separate targets for LSP and rest (#68)
By moving the different targets into separate workspaces, we avoid some
of the issues where rust-analyzer might just stop working.

By adding the two projects to .vscode/settings.json we keep the ability
to get completions, goto definitions, rename, and such operations.

This requires us to specify the dependency versions in the LSP crate, as
we can no longer refer to them by the workspace version. The positive of
this is that the WASM/LSP dependent crates are now moved to the LSP crate.