deunsafication and testing of sonic protocol
Originally sonic used unsafe casts to read and write the Header for
packets. This was "safe" as it was, but relied on the code reading and
writing to make use that it used the correct types on each end of the
protocol.

This introduces bytemuck which does the same thing, but safely. Safety
comes from bytemuck asserting at compile-time that the Header is safe
to cast to bytes, and that there are enough bytes when interpreting
the received bytes.

In addition to this, this commit also introduces proptest testing of the
sonic communication, by generating arbitray messages and sending them to
a server, and asserting that the request and response is always as
expected.