introduce type safe sonic and sonic service
The core of sonic communication now establishes a protocol that has
the request and response types determined upon creation, instead of
defered to sending and accepting. This upfront declaration makes it
harder to misuse the protocol, and more obvious what messages are
expected at each part of the protocol.

A blocker for implementing this, was in tailored request/response cycles
where a specific response type was expected based on the given request
type. Formerly, this was handeled ad-hoc, where the expected response
type was declared on the connection side to match what type the server
would send in response. This implicit coupling was brittle, and depended
opon the two call sites agreeing on what to send and when.

Based on the prior downsides, sonic::service is introduced to formalize
the request/response cycle. This is done by defining messages, which
implement a trait with an associated response type and a handler. It is
when a service connection then sends a message to a server of the same
service, the connection will statically know what response to expect,
by use of the associated type of the sent message. This provides
connections with a RPC-like interface to the service, while the service
can be expanded with new messages without encuring any additional
overhead to the call-sites of existing connections.

A new service is declared using a macro, that defined appropriate
request and response enums, used during communication. It additionally
generates the code for an ergonomic interface that uses the assiciated
response type to cover over implementation details of tagging messages
while providing a type-safe interface.

Most servers are converted to use the new sonic::service, but some still
remain at the more lowlevel sonic, which is now type-safe, since they
are used in ways incompatible with strict request/response +
predetermined types model; notably MapReduce is generic over the job,
and has non-service-like-control-flow.