`: Clone` bye bye!
By adding a new `RequestRef<'a>` for the sending part of sonic::service
we get to take references to the send messages, and thus remove the
need for copying each message sent!

This relies on `Request` and `RequestRef` being the same shape when serialized and deserialized, since we serialize the ref version but
deserialzie the non-ref version. Since both are generated by the same
macro, we can have quite high confidence in that they are the same, but
we should perhaps add some tests to ensure that this are the case. I
don't think it would be possible to exhastiavly check, but a proptest
cycling through serialize/deserialize loop could give us some confidence
that it is correct.