Safe search (#79)
* This should fix the byte/char index mixups identified in issue 77

* script to generate dataset

* naive bayes classification with tf-idf features

* Add prediction confidence to naive bayes.
we report the confidence as $log_probs[best] / sum(log_probs)$.
I'm not really sure this confidence calculation can be seen as a probability that the model has predicted the correct label, but should still give a picture of the confidence of the prediction. It's therefore named confidence and not probability.

Also, even though naive bayes is a pretty decent classifier some people on stackexchange report that it's a pretty bad probability estimator. Further tests will determine if this confidence score is actually useful.

* naive bayes benchmark

* store safe search classification in index

* search preferences page where user can control safe search settings