Fix iOS searchbar bluring and HTML escape suggestions
iOS Safari has this behavior where they will issue the blur event for
focused elements before handling onclick events. This means that
clicking the suggestions would blur the input field, remove the
`focus-within:flex` on the suggestions, and then handle the click event
which now clicks nothing.

To prevent this, we manually control the focus, and prevent the blur
if the blur event was issued by an element within the suggestions.
It thus becomes the job of the suggestion to issue the blur after
updating the query and submitting the form.