Re-write webgraph storage backend. The webgraph storage is now essentially a '(from, to) -> label' map stored in rocksdb databases. This heavily simplifies inserts and merges, since we can now insert new edges directly into the db without having to read the existing edges.
Get operations now uses a prefix iterator from rocksdb. This utilizes the fact that '[{from_bytes},0,0,0,...]' is a prefix of any '[{from_bytes},{to_bytes}]' that might have been inserted into the database.
Assuming that we use a sufficiently large read-ahead size, I think there shouldn't be a noticeable increase in IO operations for get operations and thus not noticeable performance penalty. In fact, they might be a bit faster in practice due to not having to deserialize a hashmap and from the fact that rocksdb seems to be more tuned for small key-value sizes.