Refactor entity code and add start of Wiki template rendering (#75)
* Refactor entity code

This commit is an amalgamation of many changes of varying size (that
should have been multiple commits, sorry):
- `WikiNodeExt` for convenience methods on `Node`s
- Make `Link` offsets byte-oriented rather than char-oriented
- Make `Span::add_link` compute the link offsets internally
- Add `Span::add_node` shared by the `From<[Node]>` and `EntityBuilder`
  methods
- Refactor `EntityIterator` by using the `Deref` impl on
  `quick_xml::Event`
- Move `EntityBuilder` extraction methods out to take a shared `&[Node]`
  instead of re-parsing each.
- These extraction methods have been refactored to use iterators more
  heavly and wild-cards in matching
- Add an, early-stage, `render_template` for rendering certain Wiki
  specific templates such as pronunciations
- Refactor `maybe_prettify_entity_date` and `entity_link_to_html`
- Add expect testing to `EntityBuilder` with a debug Wiki renderer

* Use `quick_xml` `name()` in `EntityIterator`

Turns out the `Deref` impl on `Event` gives all bytes and not just the
name of the event. Instead we now extract the name so we can use it in
the match arms.

* Handle `respell` templates

* Handle `Node::ParagraphBreak` during entity rendering

When a `ParagraphBreak` is encountered a `\n` is inserted. This may or may not be appropriate for the place it's rendered. In HTML I think this
will result in a since place, which would be good, but I haven't tested
just yet.

* Handle potentially empty `Node::Template` name

* Document and move `render_template` to mod scope

* Documnet `check_abstract`

* Use a snippet structure for entity rendering instead of HTML

This is an inprogress commit, that currently has bincode issues

* Build 'DisplayedEntity' on the api server.
This avoids the issues we have with tagged enums in bincode, while wanting camelCase naming for the public API. It also has the added side-benefit of keeping the internal API more clean as all 'Displayed*' now gets constructed right before they hit the public facing endpoints.

* Fix entity info spacing between links

* Change entity snippet links to underline on hover instead of color

* Include fr-IPA wiki template

---------

Co-authored-by: Mikkel Denker <mikkel@trystract.com>