Make `site:` operator respect prefix matching in path and suffix matching in domain. This allows us to search for stuff like 'site:.com', 'site:.com/en', 'site:github.io' and 'site:example.github.io'. This essentially provides much finer grained control of the 'site:' operator.
It is implemented with some tokenization tricks. Since we know that any ' ' character gets converted into %20 in the url, we can use a space to separate where the domain stops and path starts and then treat the tokens as a phrase query. In other words, 'example.com/test' becomes ['example', '.', 'com ', '/', 'test']. When the user then searches for 'site:.com/test' it will be tokenized as ['.com ', '/' , 'test']. The ' ' will ensure that we match the suffix of the domain and the prefix of the path.