Search optimizations and prevent rocksdb write stall. The searches should now be a tiny bit faster by not having to search in unecesarry fields where all the text is allready covered in other fields. This is primarily all the '*IfHomepage' fields. They are still used for scoring.
Rocksdb has a feature where they stall writes if the compactors cannot keep up. They do this to ensure adequate read performance at all times.
Since we know that there wont be any reads when we are constructing the webgraph, we can simply not stall the writers and let the compaction catch up over time after the graph has been constructed.
In practice it seems like the compaction can almost keep up anyway - it only seems to wait on 'WaitUntilFlushWouldNotStallWrites'. This still halts approximately 50% of the cpu cores while the disk is not at full capacity either.