Thesaurus widget and minor bugfixes to calculator.
The thesaurus widget is powered by a openwordnet RDF format. This should make it very easy to add more languages when we want to.

Phrase searches triggered a string function in fend which caused the calculator to show up. The y-combinator function caused a stack overflow due to the recursion. Both of these bugs have now been fixed.