Better prioritization for which domains and urls to crawl. Each domain now starts with a score of 1.0 and is added with the score of all the incoming links for that domain. A domains score is distributed amongst all the outgoing links for that domain when it is sampled.
The intuition is that if a domain has many outgoing links, each link has relatively little value whereas if a domain has few outgoing links, each link
is more important.

This score is of course not stable and depends on the order we discover and crawl urls+domains. However, I think it will work quite well
as a crawl prioritization mechanism in practice.