Make result counts optional. This allows us to short circuit the query when we have reached max_docs_considered at the query level instead of at the collector level. This should increase search performance since the searcher then doesn't have to iterate all the results in order to return a count.
In practice we probably always want to count the results when calling the api from the frontend, but it's a nice performance parameter we can flip if we need to squeeze every drop of performance from the backend at some point. We also only count the results for api requests that needs the count.