Move collector into a crate and make AsRankingWebsite require Doc
Previously there was a blanket impl for Doc, but since Doc is now
defined in another crate, this is no longer possible due to coherence.

To work around this, we require that implementors of AsRankingWebsite
also impl Doc, which is fairly straight forward if they also impl
AsRankingWebsite, which they have to.

Not as convinient, but it works!