Remove `#[async_trait]` on `sonic::Service`
For this case, `async fn` in its current form is not sufficient as we need to specify `+ Send` and that is not comming with these features. However, we can use `impl Future<...> + Send` now, and that works great, but is a little more verbose. Fortunately, this trait is never actually implemented manually, but is implemented through a macro, so nobody has to know. :)