Remove `#[async_trait]` (#119)
* Enable the `async_fn_in_trait` feature

This is momentary until the feature is stable

* Remove `async_trait` from `DatumStream`

* Remove `async_trait` from `sonic::Message` traits impls

* Remove `async_trait` from `mapreduce::Worker` trait

* Remove `#[async_trait]` on `sonic::Service`

For this case, `async fn` in its current form is not sufficient as we need to specify `+ Send` and that is not comming with these features. However, we can use `impl Future<...> + Send` now, and that works great, but is a little more verbose. Fortunately, this trait is never actually implemented manually, but is implemented through a macro, so nobody has to know. :)

* Remove `async-trait` dependency

* Change some `async fn` to `-> impl Future<...>` to make rustc happy

Having `async fn` in `pub` code is a warning, so we use `impl Trait` instead for now

* Remove `#[feature(...)]` for `async fn` and `-> impl Trait` since it's stable now