Optics LSP Chores (#145)
* Update the linked project to new location of `optics-lsp`

* Update optics-lsp license file path

It was previously a symlink to `../LICENSE.md`, but since moving the crate into `crates/` this should now be `../../LICENSE.md`.

* Bump lsp-types and serde-wasm-bindgen

* Sort crates in optics-lsp and add itertools

* Move optic-lsp token docs out to a separate file

rustfmt does not like long string literals, so these doc strings prevented most of the file from being formatted. Moving it out still mess with rustfmt, but now it is at least isolated :)

* Add token to hover docs

This blatantly steals a feature from rust-analyzer, where the hovered token is displayed as the first thing in the hover docs. It helps signalize that the documentation corresponds to the token being hovered, and also just looks nice. :)

* Refactor error formatting in optics-lsp

Instead of pushing string literals, we build up an iterator of lines and join them at the end.

Other minor refactorings are performed as well.

* change optics extension license to MIT

* release optics extension 0.0.12