Distributed in-memory key/value store for mapreduce (#181)
* [WIP] raft consensus using openraft on sonic networking

* handle rpc's on nodes

* handle get/set application requests

* dht get/set stubs that handles leader changes and retries
also improve sonic error handling. there is no need for handle to return a sonic::Result, it's better that the specific message has a Result<...> as their response as this can then be properly handled on the caller side

* join existing raft cluster

* make sure node state is consisten in case of crash -> rejoin

* ResilientConnection in sonic didn't retry requests, only connections, and was therefore a bit misleading. remove it and add a send_with_timeout_retry method to normal connection with sane defaults in .send method

* add Response::Empty to raft in order to avoid having to send back hacky Response::Set(Ok(())) for internal raft entries

* change key/value in dht to be arbitrary bytes

* dht chaos proptest

* make dht tests more reliable
in raft, writes are written to a majority quorom. if we have a cluster of 3 nodes, this means that we can only be sure that 2 of the nodes get's the data. the test might therefore fail if we are unlucky and check the node that didn't get the data yet. by having a cluster of 2 nodes instead, we can be sure that both nodes always receives all writes.

* sharded dht client