[WIP] Implemented `keybind` module to handle keyboard shortcuts (#186)
* Implemented `keybind` module to handle keyboard shortcuts

* Removal of direct DOM querying and the addition of searchbar keybindings

* Remove generics from 'keybind'
It's always used with 'Refs' as context, so there is no need to have it generic

* Revert 'Searchbar' to use simple keydown match instead of 'Keybind'
The functionality didn't work (for instance enter didn't trigger a search). It would require a lot of aditional complexity in 'Keybind' to also support the use case from searchbar. It's okay to have some code duplication if this results in a simpler solution that will therefore be more readable and mantainable long term

* Remove need to know about keyboard event in keybind callbacks
This forces us to not rely on direct manipulation of the event, but instead implement the necesarry functionality in helper methods in the different components

* forgot to remove a console.log...

---------

Co-authored-by: Mikkel Denker <mikkel@stract.com>